BugFix: Remounting
The previous commit was supposed to fix the remounting issue as I
thought it was caused by `useState` vs `useRef`, but once I started
using it, nothing changed. After looking deeper into react, the
component is remounted if its key changes. So the key is now the
location key instead of index so the key will never change, but still be
unique